# EVM: optimize stack