--- a/packages/evm/src/stack.ts
+++ b/packages/evm/src/stack.ts
@@ -1,5 +1,3 @@
-import { MAX_INTEGER_BIGINT } from '@ethereumjs/util'
-
 import { ERROR, EvmError } from './exceptions.js'
 
 /**
@@ -9,38 +7,32 @@ export class Stack {
   _store: bigint[]
   _maxHeight: number
 
+  private _len: number = 0
+
   constructor(maxHeight?: number) {
-    this._store = []
+    this._store = [] //Array(maxHeight ?? 1024) -> initializing the array makes the constructor 10x slower
     this._maxHeight = maxHeight ?? 1024
   }
 
   get length() {
-    return this._store.length
+    return this._len
   }
 
   push(value: bigint) {
-    if (typeof value !== 'bigint') {
-      throw new EvmError(ERROR.INTERNAL_ERROR)
-    }
-
-    if (value > MAX_INTEGER_BIGINT) {
-      throw new EvmError(ERROR.OUT_OF_RANGE)
-    }
-
-    if (this._store.length >= this._maxHeight) {
+    if (this._len >= this._maxHeight) {
       throw new EvmError(ERROR.STACK_OVERFLOW)
     }
 
-    this._store.push(value)
+    this._store[this._len++] = value
   }
 
   pop(): bigint {
-    if (this._store.length < 1) {
+    if (this._len < 1) {
       throw new EvmError(ERROR.STACK_UNDERFLOW)
     }
 
     // Length is checked above, so pop shouldn't return undefined
-    return this._store.pop()!
+    return this._store[--this._len]
   }
 
   /**
@@ -49,7 +41,7 @@ export class Stack {
    * @param num - Number of items to pop
    */
   popN(num: number = 1): bigint[] {
-    if (this._store.length < num) {
+    if (this._len < num) {
       throw new EvmError(ERROR.STACK_UNDERFLOW)
     }
 
@@ -57,7 +49,14 @@ export class Stack {
       return []
     }
 
-    return this._store.splice(-1 * num).reverse()
+    const arr = Array(num)
+    const cache = this._store
+
+    for (let pop = 0; pop < num; pop++) {
+      arr[pop] = cache[--this._len]
+    }
+
+    return arr
   }
 
   /**
@@ -66,14 +65,15 @@ export class Stack {
    * @throws {@link ERROR.STACK_UNDERFLOW}
    */
   peek(num: number = 1): bigint[] {
-    const peekArray: bigint[] = []
+    const peekArray: bigint[] = Array(num)
+    let start = this._len
 
-    for (let peek = 1; peek <= num; peek++) {
-      const index = this._store.length - peek
+    for (let peek = 0; peek < num; peek++) {
+      const index = --start
       if (index < 0) {
         throw new EvmError(ERROR.STACK_UNDERFLOW)
       }
-      peekArray.push(this._store[index])
+      peekArray[peek] = this._store[index]
     }
     return peekArray
   }
@@ -83,16 +83,17 @@ export class Stack {
    * @param position - Index of item from top of the stack (0-indexed)
    */
   swap(position: number) {
-    if (this._store.length <= position) {
+    if (this._len <= position) {
       throw new EvmError(ERROR.STACK_UNDERFLOW)
     }
 
-    const head = this._store.length - 1
-    const i = this._store.length - position - 1
+    const head = this._len - 1
+    const i = head - position
+    const sto = this._store
 
-    const tmp = this._store[head]
-    this._store[head] = this._store[i]
-    this._store[i] = tmp
+    const tmp = sto[head]
+    sto[head] = sto[i]
+    sto[i] = tmp
   }
 
   /**
@@ -104,11 +105,11 @@ export class Stack {
   // Nevertheless not sure if we "loose" something here?
   // Will keep commented out for now
   dup(position: number) {
-    if (this._store.length < position) {
+    if (this._len < position) {
       throw new EvmError(ERROR.STACK_UNDERFLOW)
     }
 
-    const i = this._store.length - position
-    this.push(this._store[i])
+    const i = this._len - position
+    this._store[this._len++] = this._store[i]
   }
 }