--- a/packages/evm/test/customOpcodes.spec.ts
+++ b/packages/evm/test/customOpcodes.spec.ts
@@ -40,7 +40,7 @@ describe('VM: custom opcodes', () => {
       gasLimit: BigInt(gas),
     })
     assert.ok(res.executionGasUsed === totalFee, 'successfully charged correct gas')
-    assert.ok(res.runState!.stack._store[0] === stackPush, 'successfully ran opcode logic')
+    assert.ok(res.runState!.stack.peek()[0] === stackPush, 'successfully ran opcode logic')
     assert.ok(correctOpcodeName, 'successfully set opcode name')
   })
 
@@ -97,7 +97,7 @@ describe('VM: custom opcodes', () => {
       gasLimit: BigInt(gas),
     })
     assert.ok(res.executionGasUsed === totalFee, 'successfully charged correct gas')
-    assert.ok(res.runState!.stack._store[0] === stackPush, 'successfully ran opcode logic')
+    assert.ok(res.runState!.stack.peek()[0] === stackPush, 'successfully ran opcode logic')
   })
 
   it('should pass the correct EVM options when copying the EVM', async () => {