This optimizes stack.

In particular, our previous stack kept deleting / adding items to stack. Here, the internal stack Array will only grow. 

There are two error checks removed from `stack.push`: since we only call the stack internally, we are sure: 

- The argument is always of type `bigint`
- The argument is always within `2**256-1` (`uint256`)
- Therefore, we don't have to do these error checks.

I have tested initializing the Stack with 1024 empty items, this makes the constructor 10x slower and the gains are not that good. On top of the optimizations of #2996 this optimizes the performance JUMPDEST tests from ~8.1 MGas/s to 8.5 MGas/s. (This is the test which uses this code: `0x60025B8056` -> just infinitely keep JUMPing to a JUMPDEST)

Note: target branch is also the PR of #2996.

To profile the new stack, follow the instructions in this gist: https://gist.github.com/jochem-brouwer/3969933a0eb6bb4861c1ef4e61ed3f70

Output from my side:

```
push on empty stack: Diff (neg is better): -0.377s (-41.08%)
push on almost-full-stack: Diff (neg is better): -0.261s (-29.95%)
pop on empty stack: Diff (neg is better): -0.021s (-3.48%)
pop on full-stack: Diff (neg is better): -0.028s (-4.97%)
popN 1 on stack size 1: Diff (neg is better): -0.386s (-39.98%)
popN 1 on full-stack: Diff (neg is better): -0.398s (-38.01%)
popN 8 on stack size 8: Diff (neg is better): -0.383s (-32.91%)
popN 8 on full-stack: Diff (neg is better): -0.504s (-38.91%)
peek 1 on stack size 1: Diff (neg is better): -0.136s (-17.6%)
peek 1 on full-stack: Diff (neg is better): -0.155s (-19.25%)
peek 8 on stack size 8: Diff (neg is better): -0.066s (-7.58%)
peek 8 on full-stack: Diff (neg is better): -0.065s (-7.44%)
swap 1 on stack size 2: Diff (neg is better): 0.009s (1.48%)
swap 1 on full-stack: Diff (neg is better): 0.024s (4.3%)
swap 16 on stack size 17: Diff (neg is better): 0.017s (3.07%)
swap 16 on full-stack: Diff (neg is better): 0.015s (2.61%)
dup 1 on stack size 1: Diff (neg is better): -0.449s (-45.2%)
dup 1 on almost-full-stack: Diff (neg is better): -0.428s (-42.38%)
dup 16 on stack size 16: Diff (neg is better): -0.455s (-44.69%)
dup 16 on almost-full-stack: Diff (neg is better): -0.448s (-43.78%)
```

JUMPDEST test:

```
gas: 99979000, time: 11.483665864944458, Mgas/s: 8.70619201009673
gas: 99979000, time: 11.755131164073944, Mgas/s: 8.505136914639968
gas: 99979000, time: 11.642706596970559, Mgas/s: 8.587264410323163
gas: 99979000, time: 11.971259549975395, Mgas/s: 8.351585694272705
```


Sample: 

```
Profile: JUMPDEST loop (gas: 100000000)
+== Opcodes performance - Tx 0x4766bc7b95062417f71bcd725f5331989f0c3ecc4ab75d4db52674423d699aa6 ==+
|---------------------------------------------------------------|
|      tag |   calls | ms/call | total (ms) | gas used | Mgas/s |
|     JUMP | 8331583 |       0 |   2676.627 | 66652664 | 24.902 |
|      DUP | 8331584 |       0 |   1883.657 | 24994752 | 13.269 |
| JUMPDEST | 8331584 |       0 |    1832.62 |  8331584 |  4.546 |
|     PUSH |       1 |   0.172 |      0.172 |        3 |  0.017 |
+---------------------------------------------------------------+
gas: 99979000, time: 11.971259549975395, Mgas/s: 8.351585694272705
```