@jochem-brouwer 2023-08-31
Same version as used as results in the opening post of this PR, but now the results with initializing the stack array to 1024 items:

```
push on empty stack: Diff (neg is better): -0.321s (-36.47%)
push on almost-full-stack: Diff (neg is better): -0.255s (-29.23%)
pop on empty stack: Diff (neg is better): 0.003s (0.4%)
pop on full-stack: Diff (neg is better): -0.033s (-5.63%)
popN 1 on stack size 1: Diff (neg is better): -0.389s (-40.18%)
popN 1 on full-stack: Diff (neg is better): -0.438s (-40.6%)
popN 8 on stack size 8: Diff (neg is better): -0.398s (-33.34%)
popN 8 on full-stack: Diff (neg is better): -0.498s (-38.79%)
peek 1 on stack size 1: Diff (neg is better): -0.16s (-20.83%)
peek 1 on full-stack: Diff (neg is better): -0.094s (-12.09%)
peek 8 on stack size 8: Diff (neg is better): 0.009s (0.99%)
peek 8 on full-stack: Diff (neg is better): 0.027s (3.12%)
swap 1 on stack size 2: Diff (neg is better): 0.005s (0.8%)
swap 1 on full-stack: Diff (neg is better): 0.01s (1.73%)
swap 16 on stack size 17: Diff (neg is better): 0.013s (2.33%)
swap 16 on full-stack: Diff (neg is better): 0.02s (3.5%)
dup 1 on stack size 1: Diff (neg is better): -0.426s (-42.42%)
dup 1 on almost-full-stack: Diff (neg is better): -0.441s (-42.55%)
dup 16 on stack size 16: Diff (neg is better): -0.442s (-42.58%)
dup 16 on almost-full-stack: Diff (neg is better): -0.43s (-41.41%)
```

I don't really see a big difference here :thinking: